% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sfn_operations.R
\name{delete_activity}
\alias{delete_activity}
\title{Deletes an activity}
\usage{
delete_activity(activityArn)
}
\arguments{
\item{activityArn}{The Amazon Resource Name (ARN) of the activity to delete.}
}
\description{
Deletes an activity.
}
\examples{

}
