% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sfn_operations.R
\name{describe_activity}
\alias{describe_activity}
\title{Describes an activity}
\usage{
describe_activity(
  activityArn = "string"
)
}
\arguments{
\item{activityArn}{[required] The Amazon Resource Name (ARN) of the activity to describe.}
}
\description{
Describes an activity.
}
\details{
This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.
}
\examples{

}
