% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sfn_operations.R
\name{describe_execution}
\alias{describe_execution}
\title{Describes an execution}
\usage{
describe_execution(executionArn)
}
\arguments{
\item{executionArn}{[required] The Amazon Resource Name (ARN) of the execution to describe.}
}
\description{
Describes an execution.
}
\details{
This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.
}
\section{Accepted Parameters}{
\preformatted{describe_execution(
  executionArn = "string"
)
}
}

