% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sfn_operations.R
\name{describe_state_machine}
\alias{describe_state_machine}
\title{Describes a state machine}
\usage{
describe_state_machine(
  stateMachineArn = "string"
)
}
\arguments{
\item{stateMachineArn}{[required] The Amazon Resource Name (ARN) of the state machine to describe.}
}
\description{
Describes a state machine.
}
\details{
This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.
}
\examples{

}
