% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sfn_operations.R
\name{describe_state_machine_for_execution}
\alias{describe_state_machine_for_execution}
\title{Describes the state machine associated with a specific execution}
\usage{
describe_state_machine_for_execution(executionArn)
}
\arguments{
\item{executionArn}{[required] The Amazon Resource Name (ARN) of the execution you want state machine information for.}
}
\description{
Describes the state machine associated with a specific execution.
}
\section{Accepted Parameters}{
\preformatted{describe_state_machine_for_execution(
  executionArn = "string"
)
}
}

