% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sfn_operations.R
\name{get_execution_history}
\alias{get_execution_history}
\title{Returns the history of the specified execution as a list of events}
\usage{
get_execution_history(executionArn, maxResults = NULL,
  reverseOrder = NULL, nextToken = NULL)
}
\arguments{
\item{executionArn}{[required] The Amazon Resource Name (ARN) of the execution.}

\item{maxResults}{The maximum number of results that are returned per call. You can use \code{nextToken} to obtain further pages of results. The default is 100 and the maximum allowed page size is 100. A value of 0 uses the default.

This is only an upper limit. The actual number of results returned per call might be fewer than the specified maximum.}

\item{reverseOrder}{Lists events in descending order of their \code{timeStamp}.}

\item{nextToken}{If a \code{nextToken} is returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in \code{nextToken}. Keep all other arguments unchanged.

The configured \code{maxResults} determines how many results can be returned in a single call.}
}
\description{
Returns the history of the specified execution as a list of events. By default, the results are returned in ascending order of the \code{timeStamp} of the events. Use the \code{reverseOrder} parameter to get the latest events first.
}
\details{
If a \code{nextToken} is returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in \code{nextToken}. Keep all other arguments unchanged.
}
\section{Accepted Parameters}{
\preformatted{get_execution_history(
  executionArn = "string",
  maxResults = 123,
  reverseOrder = TRUE|FALSE,
  nextToken = "string"
)
}
}

