% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sfn_operations.R
\name{list_state_machines}
\alias{list_state_machines}
\title{Lists the existing state machines}
\usage{
list_state_machines(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results that are returned per call. You can use \code{nextToken} to obtain further pages of results. The default is 100 and the maximum allowed page size is 100. A value of 0 uses the default.

This is only an upper limit. The actual number of results returned per call might be fewer than the specified maximum.}

\item{nextToken}{If a \code{nextToken} is returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in \code{nextToken}. Keep all other arguments unchanged.

The configured \code{maxResults} determines how many results can be returned in a single call.}
}
\description{
Lists the existing state machines.
}
\details{
If a \code{nextToken} is returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in \code{nextToken}. Keep all other arguments unchanged.
}
\examples{

}
