% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sfn_operations.R
\name{list_tags_for_resource}
\alias{list_tags_for_resource}
\title{List tags for a given resource}
\usage{
list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) for the Step Functions state machine or activity.}
}
\description{
List tags for a given resource.
}
\section{Accepted Parameters}{
\preformatted{list_tags_for_resource(
  resourceArn = "string"
)
}
}

