% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sfn_package.R
\docType{package}
\name{paws.sfn-package}
\alias{paws.sfn}
\alias{paws.sfn-package}
\title{paws.sfn: AWS Step Functions}
\description{
AWS Step Functions is a service that lets you
coordinate the components of distributed applications and
microservices using visual workflows.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
