% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sfn_operations.R
\name{send_task_success}
\alias{send_task_success}
\title{Used by workers to report that the task identified by the \code{taskToken} completed successfully}
\usage{
send_task_success(taskToken, output)
}
\arguments{
\item{taskToken}{The token that represents this task. Task tokens are generated by the service when the tasks are assigned to a worker (see GetActivityTaskOutput\$taskToken).}

\item{output}{The JSON output of the task.}
}
\description{
Used by workers to report that the task identified by the \code{taskToken} completed successfully.
}
\examples{

}
