% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sfn_operations.R
\name{start_execution}
\alias{start_execution}
\title{Starts a state machine execution}
\usage{
start_execution(
  stateMachineArn = "string",
  name = "string",
  input = "string"
)
}
\arguments{
\item{stateMachineArn}{[required] The Amazon Resource Name (ARN) of the state machine to execute.}

\item{name}{The name of the execution. This name must be unique for your AWS account and region for 90 days. For more information, see \href{http://docs.aws.amazon.com/step-functions/latest/dg/limits.html#service-limits-state-machine-executions}{Limits Related to State Machine Executions} in the \emph{AWS Step Functions Developer Guide}.

A name must \emph{not} contain:
\itemize{
\item whitespace
\item brackets \code{< > { } [ ]}
\item wildcard characters \code{? *}
\item special characters \code{\" # % \ ^ | ~ ` $ & , ; : /}
\item control characters (\code{U+0000-001F}, \code{U+007F-009F})
}

[ ]: R:%20}

\item{input}{The string that contains the JSON input data for the execution, for example:

\code{"input": "{\"first_name\" : \"test\"}"}

If you don't include any JSON input data, you still must include the two braces, for example: \code{"input": "{}"}}
}
\description{
Starts a state machine execution.
}
\details{
\code{StartExecution} is idempotent. If \code{StartExecution} is called with the same name and input as a running execution, the call will succeed and return the same response as the original request. If the execution is closed or if the input is different, it will return a 400 \code{ExecutionAlreadyExists} error. Names can be reused after 90 days.
}
\examples{

}
