% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sfn_operations.R
\name{untag_resource}
\alias{untag_resource}
\title{Remove a tag from a Step Functions resource}
\usage{
untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) for the Step Functions state machine or activity.}

\item{tagKeys}{[required] The list of tags to remove from the resource.}
}
\description{
Remove a tag from a Step Functions resource
}
\section{Accepted Parameters}{
\preformatted{untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

