% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sfn_operations.R
\name{create_state_machine}
\alias{create_state_machine}
\title{Creates a state machine}
\usage{
create_state_machine(name, definition, roleArn)
}
\arguments{
\item{name}{The name of the state machine. This name must be unique for your AWS account and region for 90 days. For more information, see \href{http://docs.aws.amazon.com/step-functions/latest/dg/limits.html#service-limits-state-machine-executions}{Limits Related to State Machine Executions} in the \emph{AWS Step Functions Developer Guide}.

A name must \emph{not} contain:
\itemize{
\item whitespace
\item brackets \code{< > { } [ ]}
\item wildcard characters \code{? *}
\item special characters \code{\" # % \ ^ | ~ ` $ & , ; : /}
\item control characters (\code{U+0000-001F}, \code{U+007F-009F})
}

[ ]: R:%20}

\item{definition}{The Amazon States Language definition of the state machine.}

\item{roleArn}{The Amazon Resource Name (ARN) of the IAM role to use for this state machine.}
}
\description{
Creates a state machine. A state machine consists of a collection of states that can do work (\code{Task} states), determine to which states to transition next (\code{Choice} states), stop an execution with an error (\code{Fail} states), and so on. State machines are specified using a JSON-based, structured language.
}
\examples{

}
