% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sfn_operations.R
\name{send_task_failure}
\alias{send_task_failure}
\title{Used by workers to report that the task identified by the \code{taskToken} failed}
\usage{
send_task_failure(taskToken, error = NULL, cause = NULL)
}
\arguments{
\item{taskToken}{The token that represents this task. Task tokens are generated by the service when the tasks are assigned to a worker (see GetActivityTask::taskToken).}

\item{error}{An arbitrary error code that identifies the cause of the failure.}

\item{cause}{A more detailed explanation of the cause of the failure.}
}
\description{
Used by workers to report that the task identified by the \code{taskToken} failed.
}
\examples{

}
