% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sfn_operations.R
\name{stop_execution}
\alias{stop_execution}
\title{Stops an execution}
\usage{
stop_execution(executionArn, error = NULL, cause = NULL)
}
\arguments{
\item{executionArn}{[required] The Amazon Resource Name (ARN) of the execution to stop.}

\item{error}{The error code of the failure.}

\item{cause}{A more detailed explanation of the cause of the failure.}
}
\description{
Stops an execution.
}
\section{Accepted Parameters}{
\preformatted{stop_execution(
  executionArn = "string",
  error = "string",
  cause = "string"
)
}
}

